Overlay, Format80 compression
- Implemented Overlay parsing completely
- Implemented Format80 compression (uses command 1 to start and end a
chunk, this is how WW seems to have used it, too for some reason)
- Minor changes

+++ Map_Editor_BaseCode/Map_Editor_BaseCode/Editor.Engine/Types/Overlay.cpp
@@ -7,7 +7,7 @@ Overlay::Overlay()
 {
 }
 
-Overlay::Overlay(int x, int y, int overlayIndex, int overlayFrame)
+Overlay::Overlay(int x, int y, byte overlayIndex, byte overlayFrame)
 :overlayIndex(overlayIndex), overlayFrame(overlayFrame)
 {
 	loc.x = x;
